chore: Move layout to separate component
