+++ src/routes/__root.tsx
@@ -1,33 +1,5 @@
-import {
-    CalendarMonth,
-    Menu as MenuIcon,
-    MenuOpen,
-    TrendingUp,
-} from "@mui/icons-material";
-import {
-    AppBar,
-    Box,
-    Divider,
-    Drawer,
-    drawerClasses,
-    IconButton,
-    List,
-    ListItem,
-    ListItemButton,
-    ListItemIcon,
-    ListItemText,
-    Menu,
-    MenuItem,
-    Stack,
-    Toolbar,
-    Typography,
-} from "@mui/material";
-import {
-    createLink,
-    createRootRoute,
-    Outlet,
-    useLocation,
-} from "@tanstack/react-router";
+import { Layout } from "#components/Layout.tsx";
+import { createRootRoute, Outlet } from "@tanstack/react-router";
 import React, { Suspense } from "react";
 
 const TanStackRouterDevtools =
@@ -40,188 +12,16 @@ const TanStackRouterDevtools =
               })),
           );
 
-const RouterLinkItemButton = createLink(ListItemButton);
-const RouterMenuItem = createLink(MenuItem);
-
-const APP_BAR_HEIGHT_PX = "64px";
-
 const RouteComponent: React.FC = () => {
-    const location = useLocation();
-    const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-    const menuOpen = Boolean(anchorEl);
-    const handleClickMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleCloseMenu = () => {
-        setAnchorEl(null);
-    };
-
     return (
-        // Layout inspired by https://github.com/mui/material-ui/tree/v6.4.1/docs/data/material/getting-started/templates/dashboard
-        <Box sx={{ display: "flex" }}>
-            <AppBar
-                position="fixed"
-                sx={{
-                    display: { xs: "auto", md: "none" },
-                    boxShadow: 0,
-                    bgcolor: "background.paper",
-                    borderBottom: "1px solid",
-                    borderColor: "divider",
-                    height: APP_BAR_HEIGHT_PX,
-                }}
-            >
-                <Toolbar variant="regular">
-                    <Stack
-                        direction="row"
-                        spacing={2}
-                        alignItems="center"
-                        justifyContent="flex-end"
-                        width="100%"
-                    >
-                        <IconButton onClick={handleClickMenu}>
-                            {menuOpen ? <MenuOpen /> : <MenuIcon />}
-                        </IconButton>
-                        <Menu
-                            open={menuOpen}
-                            anchorEl={anchorEl}
-                            onClose={handleCloseMenu}
-                        >
-                            <RouterMenuItem
-                                to="/session"
-                                search={{
-                                    uuid: "a937646b-f115-44c3-8dbf-9ae4a65669a0",
-                                    timeInterval: {
-                                        type: "lastXDays",
-                                        end: new Date(
-                                            "2025-01-15T22:30:00+01:00",
-                                        ),
-                                    },
-                                    gamemode: "overall",
-                                    stat: "fkdr",
-                                    variantSelection: "both",
-                                }}
-                            >
-                                <ListItemIcon>
-                                    <TrendingUp />
-                                </ListItemIcon>
-                                <ListItemText primary="Session stats" />
-                            </RouterMenuItem>
-                            <RouterMenuItem
-                                to="/history/compare"
-                                search={{
-                                    uuids: [
-                                        "a937646b-f115-44c3-8dbf-9ae4a65669a0", // Skydeath
-                                        "ac04f297-f74c-44de-a24e-0083936ac59a", // USBB
-                                        "062c373b-28de-4ec0-ab2c-0114e59e36ce", // Skydeaf
-                                        "3d58a2de-3831-4a17-a305-67258295f81e", // iCiara
-                                        "6bc1dd0f-f351-4c3d-b6cc-262e55b6e7aa", // HardcoreLizard
-                                    ],
-                                    start: new Date(
-                                        "2025-01-15T19:55:00+01:00",
-                                    ),
-                                    end: new Date("2025-01-15T22:30:00+01:00"),
-                                    limit: 100,
-                                    stats: ["stars", "finalKills"],
-                                    gamemodes: ["overall"],
-                                    variants: ["session"],
-                                }}
-                            >
-                                <ListItemIcon>
-                                    <CalendarMonth />
-                                </ListItemIcon>
-                                <ListItemText primary="Compare history" />
-                            </RouterMenuItem>
-                        </Menu>
-                    </Stack>
-                </Toolbar>
-            </AppBar>
-            <Drawer
-                variant="permanent"
-                sx={{
-                    width: 240,
-                    flexShrink: 0,
-                    boxSizing: "border-box",
-                    mt: 10,
-                    display: { xs: "none", md: "block" },
-                    [`& .${drawerClasses.paper}`]: {
-                        backgroundColor: "background.paper",
-                        width: 240,
-                        boxSizing: "border-box",
-                    },
-                }}
-            >
-                {/*TODO: Link to "home"*/}
-                <Stack
-                    direction="row"
-                    padding={1.5}
-                    gap={1}
-                    alignItems="center"
-                >
-                    <img src="/who.png" alt="Prism Overlay logo" width="40" />
-                    <Typography variant="h6">Prism Overlay</Typography>
-                </Stack>
-                <Divider />
-                <List dense component="menu">
-                    <ListItem disablePadding>
-                        <RouterLinkItemButton
-                            selected={location.pathname === "/session"}
-                            to="/session"
-                            search={{
-                                uuid: "a937646b-f115-44c3-8dbf-9ae4a65669a0",
-                                timeInterval: {
-                                    type: "lastXDays",
-                                    end: new Date("2025-01-15T22:30:00+01:00"),
-                                },
-                                gamemode: "overall",
-                                stat: "fkdr",
-                                variantSelection: "both",
-                            }}
-                        >
-                            <ListItemIcon>
-                                <TrendingUp />
-                            </ListItemIcon>
-                            <ListItemText primary="Session stats" />
-                        </RouterLinkItemButton>
-                    </ListItem>
-                    <ListItem disablePadding>
-                        <RouterLinkItemButton
-                            selected={location.pathname === "/history/compare"}
-                            to="/history/compare"
-                            search={{
-                                uuids: [
-                                    "a937646b-f115-44c3-8dbf-9ae4a65669a0", // Skydeath
-                                    "ac04f297-f74c-44de-a24e-0083936ac59a", // USBB
-                                    "062c373b-28de-4ec0-ab2c-0114e59e36ce", // Skydeaf
-                                    "3d58a2de-3831-4a17-a305-67258295f81e", // iCiara
-                                    "6bc1dd0f-f351-4c3d-b6cc-262e55b6e7aa", // HardcoreLizard
-                                ],
-                                start: new Date("2025-01-15T19:55:00+01:00"),
-                                end: new Date("2025-01-15T22:30:00+01:00"),
-                                limit: 100,
-                                stats: ["stars", "finalKills"],
-                                gamemodes: ["overall"],
-                                variants: ["session"],
-                            }}
-                        >
-                            <ListItemIcon>
-                                <CalendarMonth />
-                            </ListItemIcon>
-                            <ListItemText primary="Compare history" />
-                        </RouterLinkItemButton>
-                    </ListItem>
-                </List>
-            </Drawer>
-            <Stack
-                component="main"
-                sx={{ marginTop: { xs: APP_BAR_HEIGHT_PX, md: 0 } }}
-                padding={1}
-            >
+        <>
+            <Layout>
                 <Outlet />
-            </Stack>
+            </Layout>
             <Suspense>
                 <TanStackRouterDevtools />
             </Suspense>
-        </Box>
+        </>
     );
 };
